feat(redis): pair a standby connection for idempotent read failover
