+++ server/src/external/redis/initRedisV2.ts
@@ -17,19 +17,16 @@ const dragonflyUrl = rawDragonflyUrl
 
 export const hasRedisV2Config = Boolean(dragonflyUrl);
 
-export const redisV2: Redis = dragonflyUrl
-	? createStandbyRedisConnection({
-			cacheUrl: dragonflyUrl,
-			region: `${currentRegion}:v2`,
-			redisType: "subject-primary",
-			commandTimeout: REDIS_V2_COMMAND_TIMEOUT_MS,
-		})
-	: createRedisConnection({
-			cacheUrl: "",
-			region: `${currentRegion}:v2`,
-			redisType: "subject-primary",
-			commandTimeout: REDIS_V2_COMMAND_TIMEOUT_MS,
-		});
+const createV2Connection = dragonflyUrl
+	? createStandbyRedisConnection
+	: createRedisConnection;
+
+export const redisV2: Redis = createV2Connection({
+	cacheUrl: dragonflyUrl || "",
+	region: `${currentRegion}:v2`,
+	redisType: "subject-primary",
+	commandTimeout: REDIS_V2_COMMAND_TIMEOUT_MS,
+});
 
 const alternateInstanceUrls: Partial<Record<RedisV2InstanceName, string>> = {
 	upstash: process.env.CACHE_V2_UPSTASH_URL?.trim() || undefined,
