+++ server/src/honoUtils/handleHealthCheck.ts
@@ -2,6 +2,10 @@ import type { Context } from "hono";
 import { logger } from "@/external/logtail/logtailUtils.js";
 import { getMiscRedis } from "@/external/redis/initRedis.js";
 import { hasRedisV2Config, redisV2 } from "@/external/redis/initRedisV2.js";
+import {
+	describeRedisWithStandby,
+	isRedisReadyWithStandby,
+} from "@/external/redis/initUtils/standbyRedis.js";
 import type { HonoEnv } from "./HonoEnv";
 import { evaluateStartupGate } from "./startupGate.js";
 
@@ -27,14 +31,14 @@ const tryLatchStartupReady = () => {
 
 	const { ready, reason } = evaluateStartupGate({
 		redisReady: miscRedisStatus === "ready",
-		redisV2Ready: !hasRedisV2Config || redisV2.status === "ready",
+		redisV2Ready: !hasRedisV2Config || isRedisReadyWithStandby(redisV2),
 		elapsedMs: Date.now() - startedAt,
 	});
 	if (!ready) return;
 	startupReady = true;
 	logger.info(`[health-check] startup gate latched (${reason})`, {
 		redis_status: miscRedisStatus,
-		redis_v2_status: redisV2.status,
+		redis_v2_status: describeRedisWithStandby(redisV2),
 		has_redis_v2_config: hasRedisV2Config,
 	});
 };
