+++ server/src/internal/customers/cache/fullSubject/actions/invalidate/invalidateSharedBalanceFields.ts
@@ -1,5 +1,6 @@
 import type { SubjectBalance, UsageWindow } from "@autumn/shared";
 import type { Redis } from "ioredis";
+import { isRedisReadyWithStandby } from "@/external/redis/initUtils/standbyRedis.js";
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 import { flushSubjectBalancesToDb } from "@/internal/balances/utils/sync/flushSubjectBalancesToDb.js";
 import type { UsageWindowUpdate } from "@/internal/balances/utils/types/usageWindowUpdate.js";
@@ -43,7 +44,7 @@ export const invalidateSharedBalanceFields = async ({
 	flushBalances?: boolean;
 }): Promise<void> => {
 	const { org, env } = ctx;
-	if (!customerId || redisV2.status !== "ready") return;
+	if (!customerId || !isRedisReadyWithStandby(redisV2)) return;
 
 	const subjectKey = buildFullSubjectKey({ orgId: org.id, env, customerId });
